Mocking out card struct.

+++ tarot/src/main.rs
@@ -6,7 +6,7 @@ fn main() {
     let tarot_cards = game::MajorArcana::new();
 
     for o in tarot_cards.sorted_keys() {
-        println!("({}) {}", o, tarot_cards.card_title(*o as i32));
+        println!("({}) {}", o, tarot_cards.card_title(*o));
     }
     println!("({}) {}", 0, tarot_cards.card_title(0))
 }
